// apps/mobile/src/types/shared.types.ts
export enum MealType {
  BREAKFAST = 'BREAKFAST',
  LUNCH = 'LUNCH',
  DINNER = 'DINNER',
  SNACK = 'SNACK',
  OTHER = 'OTHER',
}

export enum FoodType {
  VEGETABLE = 'VEGETABLE',
  FRUIT = 'FRUIT',
  GRAIN = 'GRAIN',
  DESSERT = 'DESSERT',
  BEVERAGE = 'BEVERAGE',
  DAIRY = 'DAIRY',
  FATS = 'FATS',
  NUTS = 'NUTS',
  MEAT = 'MEAT',
  MEAL = 'MEAL',
}

export enum PortionUnit {
  TEASPOON = 'TEASPOON',
  TABLESPOON = 'TABLESPOON',
  CUP = 'CUP',
  MILLILITER = 'MILLILITER',
  LITER = 'LITER',
  PINT = 'PINT',
  QUART = 'QUART',
  GALLON = 'GALLON',

  GRAM = 'GRAM',
  KILOGRAM = 'KILOGRAM',
  OUNCE = 'OUNCE',
  POUND = 'POUND',

  PINCH = 'PINCH',
  DASH = 'DASH',
  PIECE = 'PIECE',
  SLICE = 'SLICE',
  SERVING = 'SERVING',
}

export enum NutritionUnit {
  g = 'g',
  kg = 'kg',
  mg = 'mg',
  mcg = 'mcg',

  L = 'L',
  mL = 'mL',
  mcL = 'mcL',

  cal = 'cal',
  kcal = 'kcal',
  kJ = 'kJ',

  oz = 'oz',
  lb = 'lb',
}